SD-1231 Endringer for å håndtere kontrollert stopping av POD.

+++ src/main/kotlin/no/nav/soknad/arkivering/soknadsarkiverer/service/ArchiverService.kt
@@ -5,6 +5,9 @@ import no.nav.soknad.arkivering.avroschemas.EventTypes.*
 import no.nav.soknad.arkivering.avroschemas.ProcessingEvent
 import no.nav.soknad.arkivering.avroschemas.Soknadarkivschema
 import no.nav.soknad.arkivering.soknadsarkiverer.arkivservice.JournalpostClientInterface
+import no.nav.soknad.arkivering.soknadsarkiverer.config.AppConfiguration
+import no.nav.soknad.arkivering.soknadsarkiverer.config.busyDec
+import no.nav.soknad.arkivering.soknadsarkiverer.config.busyInc
 import no.nav.soknad.arkivering.soknadsarkiverer.kafka.KafkaPublisher
 import no.nav.soknad.arkivering.soknadsarkiverer.service.fileservice.FileserviceInterface
 import org.slf4j.LoggerFactory
@@ -13,7 +16,8 @@ import java.io.PrintWriter
 import java.io.StringWriter
 
 @Service
-class ArchiverService(private val filestorageService: FileserviceInterface,
+class ArchiverService(private val appConfiguration: AppConfiguration,
+											private val filestorageService: FileserviceInterface,
 											private val journalpostClient: JournalpostClientInterface,
 											private val kafkaPublisher: KafkaPublisher) {
 	private val logger = LoggerFactory.getLogger(javaClass)
@@ -24,21 +28,22 @@ class ArchiverService(private val filestorageService: FileserviceInterface,
 
 			val files = filestorageService.getFilesFromFilestorage(key, data)
 
-			val journalpostId = journalpostClient.opprettJournalpost(key, data, files)
-			logger.info("${key}: Opprettet journalpostId=${journalpostId} for behandlingsid=${data.getBehandlingsid()}")
-			createProcessingEvent(key, ARCHIVED)
+			if (busyInc(appConfiguration)) {
+				val journalpostId = journalpostClient.opprettJournalpost(key, data, files)
+				logger.info("${key}: Opprettet journalpostId=${journalpostId} for behandlingsid=${data.getBehandlingsid()}")
+				createProcessingEvent(key, ARCHIVED)
+				busyDec(appConfiguration)
 
-			filestorageService.deleteFilesFromFilestorage(key, data)
-			createProcessingEvent(key, FINISHED)
-			createMessage(key, "ok")
+				filestorageService.deleteFilesFromFilestorage(key, data)
+				createProcessingEvent(key, FINISHED)
+				createMessage(key, "ok")
+			}
 
 		} catch (e: Exception) {
-
 			createMessage(key, createExceptionMessage(e))
 			throw e
 		}
 	}
-
 
 	private fun createProcessingEvent(key: String, type: EventTypes) {
 		kafkaPublisher.putProcessingEventOnTopic(key, ProcessingEvent(type))
